+++ Assets/Animation/Obstacle_Animation/FogBoxMove.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d50cd655350de5249b70814ebd544c3f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
