+++ Assets/Animation/Obstacle_Animation/FogSphereGrow.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b9a8b4fe5e34d94caecf4a921aa2197
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
