+++ Assets/Animation/Obstacle_Animation/GrowingFogSphere.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a96540203dbf702429809164b046ca15
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
