+++ Assets/Assets/Prefabs/World_Prefabs/Grassland_Prefab/Bush1.prefab
@@ -62,7 +62,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: 3ee9bfa228d9fa648917e1619f4aca68, type: 2}
+  - {fileID: 2100000, guid: fa7bf75f2c94daf4195d4c84ce55d3e9, type: 2}
   - {fileID: 2100000, guid: faa0a0508f7463d42b59a26b86f96335, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
