+++ Assets/Assets/Prefabs/World_Prefabs/Grassland_Prefab/Cattail Patch_01.prefab
@@ -61,7 +61,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: b20cb3153f8f6ff498c4c1af0ce171a9, type: 2}
+  - {fileID: 2100000, guid: fa7bf75f2c94daf4195d4c84ce55d3e9, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
