+++ Assets/Assets/Prefabs/World_Prefabs/Grassland_Prefab/REDFlowerGrasslandLrg.prefab
@@ -27,7 +27,7 @@ Transform:
   m_GameObject: {fileID: 4087569867757658422}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -1.93, y: 1.31, z: 3.64}
-  m_LocalScale: {x: 0.23, y: 0.23, z: 0.23}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 1
   m_Children: []
   m_Father: {fileID: 0}
