+++ Assets/Materials/World Materials/Terrain Materials/GrasslandTerrain2.mat
@@ -133,7 +133,7 @@ Material:
     - _Color: {r: 1, g: 1, b: 1, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _SpecColor: {r: 0.19999996, g: 0.19999996, b: 0.19999996, a: 1}
-    - _Test_Color: {r: 0.8, g: 0.7844155, b: 0.13333331, a: 0}
+    - _Test_Color: {r: 0.8, g: 0.6927453, b: 0.13333331, a: 0}
     - _test_col: {r: 0, g: 0, b: 0, a: 0}
   m_BuildTextureStacks: []
 --- !u!114 &8252821223593920753
