--- Assets/Scripts/Obstacle_Scripts/FallingTreeBehavior.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f092a6b7d4215824a97420ab026420d1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
