+++ Assets/Scripts/Obstacle_Scripts/GrassLandObstacleScripts/FallingTreeBehavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f092a6b7d4215824a97420ab026420d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
