--- Assets/Shaders/Grassland Shaders/GrasslandGrassMat.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: faa0a0508f7463d42b59a26b86f96335
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
