+++ Assets/Shaders/Grassland Shaders/GrasslandGrassMat2.mat
@@ -65,8 +65,8 @@ Material:
     - _Smoothness: 0
     - _Speed: 1
     m_Colors:
-    - _Bottom_Color: {r: 0.75294125, g: 0.7607844, b: 0.16470589, a: 1}
+    - _Bottom_Color: {r: 0.7568628, g: 0.65882355, b: 0.16470589, a: 1}
     - _Color: {r: 0.7504202, g: 1, b: 0, a: 0}
     - _Color_2: {r: 0.9051054, g: 1, b: 0.36320752, a: 0}
-    - _Top_Color: {r: 0.84666467, g: 0.8584906, b: 0.19032574, a: 1}
+    - _Top_Color: {r: 0.85882354, g: 0.53817976, b: 0.19215685, a: 1}
   m_BuildTextureStacks: []
