+++ ProjectSettings/Packages/com.unity.polybrush/Settings.json
@@ -114,7 +114,7 @@
             {
                 "type": "UnityEditor.Polybrush.PrefabLoadout, Unity.Polybrush.Editor, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null",
                 "key": "ScatteringEditor.userLoadout",
-                "value": "{\"m_Value\":{\"infos\":[{\"palette\":{\"fileID\":11400000,\"guid\":\"63c2baaa9bdaa5e46a286d06c1d0aa22\",\"type\":2},\"prefab\":{\"fileID\":2760805095522722099,\"guid\":\"3addaf6c0b57fc74c8cc9ddef0126fc0\",\"type\":3}},{\"palette\":{\"fileID\":11400000,\"guid\":\"63c2baaa9bdaa5e46a286d06c1d0aa22\",\"type\":2},\"prefab\":{\"fileID\":3660816877972815924,\"guid\":\"a9ca6761bd891d04888468882c7cacf7\",\"type\":3}},{\"palette\":{\"fileID\":11400000,\"guid\":\"63c2baaa9bdaa5e46a286d06c1d0aa22\",\"type\":2},\"prefab\":{\"fileID\":3660816877972815924,\"guid\":\"fdbac08c06334f048822a2e1a0c9feec\",\"type\":3}}]}}"
+                "value": "{\"m_Value\":{\"infos\":[{\"palette\":{\"fileID\":11400000,\"guid\":\"63c2baaa9bdaa5e46a286d06c1d0aa22\",\"type\":2},\"prefab\":{\"fileID\":5795014270389188853,\"guid\":\"60c803b3e6224d54ebbc69d158a4ef87\",\"type\":3}}]}}"
             }
         ]
     }
